+++ main.c
@@ -6,7 +6,7 @@
 
 int main(int argc, char **argv) {
     if (argc == 2 && strcmp(argv[1], "--version") == 0) {
-        puts("SianLang 0.5.0");
+        puts("SianLang 0.5.1");
         return 0;
     }
     if (argc != 2) {